+++ stats/histogrammarker.h
@@ -5,12 +5,11 @@
 #include "chartitem.h"
 
 class StatsAxis;
-class StatsView;
 
 // A line marking median or mean in histograms
 class HistogramMarker : public ChartLineItem {
 public:
-	HistogramMarker(StatsView &view, double val, bool horizontal, QColor color, StatsAxis *xAxis, StatsAxis *yAxis);
+	HistogramMarker(ChartView &view, double val, bool horizontal, QColor color, StatsAxis *xAxis, StatsAxis *yAxis);
 	void updatePosition();
 private:
 	StatsAxis *xAxis, *yAxis;
