+++ stats/informationbox.h
@@ -10,11 +10,12 @@
 #include <memory>
 
 struct dive;
-class StatsView;
+class ChartView;
+class StatsTheme;
 
 // Information window showing data of highlighted dive
 struct InformationBox : ChartRectItem {
-	InformationBox(StatsView &);
+	InformationBox(ChartView &, const StatsTheme &theme);
 	void setText(const std::vector<QString> &text, QPointF pos);
 	void setPos(QPointF pos);
 	int recommendedMaxLines() const;
