+++ stats/quartilemarker.h
@@ -5,11 +5,10 @@
 #include "chartitem.h"
 
 class StatsAxis;
-class StatsView;
 
 class QuartileMarker : public ChartLineItem {
 public:
-	QuartileMarker(StatsView &view, double pos, double value, StatsAxis *xAxis, StatsAxis *yAxis);
+	QuartileMarker(ChartView &view, const StatsTheme &theme, double pos, double value, StatsAxis *xAxis, StatsAxis *yAxis);
 	~QuartileMarker();
 	void updatePosition();
 private:
