+++ stats/regressionitem.h
@@ -6,7 +6,6 @@
 
 class StatsAxis;
 class StatsTheme;
-class StatsView;
 
 struct regression_data {
 	double a,b;
@@ -16,7 +15,7 @@ struct regression_data {
 
 class RegressionItem : public ChartPixmapItem {
 public:
-	RegressionItem(StatsView &view, regression_data data, StatsAxis *xAxis, StatsAxis *yAxis);
+	RegressionItem(ChartView &view, const StatsTheme &theme, regression_data data, StatsAxis *xAxis, StatsAxis *yAxis);
 	~RegressionItem();
 	void updatePosition();
 	void setFeatures(bool regression, bool confidence);
