+++ stats/scatterseries.h
@@ -6,6 +6,7 @@
 
 #include "statshelper.h"
 #include "statsseries.h"
+#include "qt-quick/chartitem_ptr.h"
 
 #include <memory>
 #include <vector>
